+++ src/components/feed/sub/FeedSubMenu.tsx
@@ -3,7 +3,6 @@
 import { EllipsisHorizontalIcon } from "@heroicons/react/24/solid";
 import React, { useEffect, useRef, useState } from "react";
 import { useRouter } from "next/navigation";
-import { useDeleteFeed } from "@/lib/feed/hooks";
 import { DeleteConfirmModal } from "@/components/modal/check";
 
 interface FeedSubMenuProps {
@@ -21,7 +20,6 @@ export default function FeedSubMenu({
   const [isModalVisible, setIsModalVisible] = useState(false);
   const menuRef = useRef<HTMLDivElement>(null);
   const router = useRouter();
-  const { mutateAsync: deleteFeedMutate } = useDeleteFeed();
 
   const toggleMenu = () => {
     setIsVisible(!isVisible);
@@ -36,19 +34,9 @@ export default function FeedSubMenu({
     setIsModalVisible(false);
   };
 
-  const handleDelete = async () => {
-    try {
-      await deleteFeedMutate({ feedId, imagePath });
-      console.log("삭제되었습니다.");
-      setIsModalVisible(false);
-    } catch (error) {
-      console.error("삭제 중 오류 발생:", error);
-    }
-  };
-
   const handleEdit = () => {
     router.push(`/feed/edit/${feedId}`);
-    setIsModalVisible(false);
+    setIsVisible(false);
   };
 
   useEffect(() => {
@@ -93,7 +81,11 @@ export default function FeedSubMenu({
         </div>
       )}
       {isModalVisible && (
-        <DeleteConfirmModal onClose={closeModal} onDelete={handleDelete} />
+        <DeleteConfirmModal
+          feedId={feedId}
+          imagePath={imagePath}
+          onClose={closeModal}
+        />
       )}
     </div>
   );
